feat: extract env package
